+++ R/MTDest.R
@@ -56,9 +56,7 @@
 #' fitted model (see \code{\link{oscillation}}).
 #'
 #' @seealso
-#' Methods for fitted objects: \code{\link{print.MTDest}},
-#' \code{\link{summary.MTDest}}, \code{\link{print.summary.MTDest}},
-#' \code{\link{coef.MTDest}}, \code{\link{logLik.MTDest}}.
+#' Methods for fitted objects: \code{\link{MTDest-methods}}.
 #' Model constructor and related utilities: \code{\link{MTDmodel}},
 #' \code{\link{oscillation}}.
 #' Coercion helper: \code{\link{as.MTD}}